Move `CreateRegistrationNotification` command
