+++ wcfsetup/install/files/lib/form/RegisterActivationForm.class.php
@@ -12,7 +12,6 @@ use wcf\system\form\builder\field\TextFormField;
 use wcf\system\form\builder\field\validation\FormFieldValidationError;
 use wcf\system\form\builder\field\validation\FormFieldValidator;
 use wcf\system\request\LinkHandler;
-use wcf\system\user\command\CreateRegistrationNotification;
 use wcf\system\WCF;
 use wcf\util\HeaderUtil;
 use wcf\util\HtmlString;
@@ -151,7 +150,7 @@ final class RegisterActivationForm extends AbstractFormBuilderForm
         }
 
         // User must be reloaded to get the correct activation status.
-        $command = new CreateRegistrationNotification(new User($this->user->userID));
+        $command = new \wcf\command\user\CreateRegistrationNotification(new User($this->user->userID));
         $command();
 
         HeaderUtil::delayedRedirect(LinkHandler::getInstance()->getLink(), $redirectText, 10, 'success', true);
